Factor out stratified samplers. Improve documentation, fix minor issues
